move AvailableInLanguage Composable to it's own source file
